+++ cli/section_cli.go
@@ -2,8 +2,6 @@ package cli
 
 import (
 	"sort"
-
-	"go.bbkane.com/warg/completion"
 )
 
 // SectionMap holds Sections - used by other Sections
@@ -95,24 +93,4 @@ func (s *SectionIterator) Next() FlatSection {
 // HasNext is used in BreadthFirst. See BreadthFirst docs
 func (s *SectionIterator) HasNext() bool {
 	return len(s.queue) > 0
-}
-
-func (s *Section) CompletionCandidates() (completion.CompletionCandidates, error) {
-	ret := completion.CompletionCandidates{
-		Type:   completion.CompletionType_ValueDescription,
-		Values: []completion.CompletionCandidate{},
-	}
-	for _, name := range s.Commands.SortedNames() {
-		ret.Values = append(ret.Values, completion.CompletionCandidate{
-			Name:        string(name),
-			Description: string(s.Commands[name].HelpShort),
-		})
-	}
-	for _, name := range s.Sections.SortedNames() {
-		ret.Values = append(ret.Values, completion.CompletionCandidate{
-			Name:        string(name),
-			Description: string(s.Sections[name].HelpShort),
-		})
-	}
-	return ret, nil
 }